moved like logic for some pages to a hook
